blin

+++ libft/ft_strlen.c
@@ -10,9 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/libft.h"
+#include "../includes/cub3d.h"
 
-size_t		ft_strlen(const char *s)
+int		ft_strlen(const char *s)
 {
 	int i;
