+++ libft/ft_substr.c
@@ -10,13 +10,13 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/libft.h"
+#include "../includes/cub3d.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char		*substr;
 	char const	*news;
-	size_t		i;
+	int			i;
 
 	news = s;
 	substr = NULL;
